+++ material_symbols.ts
@@ -1,7 +1,8 @@
 import material, {
   type Options as MaterialOptions,
   type Type,
-} from "./catalogs/material.ts";
+  type Weight,
+} from "./catalogs/material_symbols.ts";
 
 export interface Options extends MaterialOptions {
   /** Filter name */
@@ -9,7 +10,8 @@ export interface Options extends MaterialOptions {
 }
 
 export const defaults: Options = {
-  defaultType: "filled",
+  defaultType: "outlined",
+  defaultWeight: 400,
   name: "material",
 };
 
@@ -18,12 +20,12 @@ export default function plugin(userOptions?: Options) {
   const catalog = material(options);
 
   return (site: Lume.Site) => {
-    site.filter(options.name, (name: string, type?: Type) => {
+    site.filter(options.name, (name: string, weight?: Weight, type?: Type) => {
       if (!name) {
         return "";
       }
 
-      return catalog.get(name, type);
+      return catalog.get(name, weight, type);
     }, true);
   };
 }