+++ test.ts
@@ -3,6 +3,7 @@ import phosphor from "./catalogs/phosphor.ts";
 import simpleIcons from "./catalogs/simpleicons.ts";
 import heroIcons from "./catalogs/heroicons.ts";
 import material from "./catalogs/material.ts";
+import materialSymbols from "./catalogs/material_symbols.ts";
 
 Deno.test("Phosphor", async () => {
   const catalog = phosphor();
@@ -52,5 +53,14 @@ Deno.test("Material", async () => {
   assertEquals(
     await catalog.get("10k"),
     `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24"><path d="M10 10.5h1.5v3H10zM19 3H5c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h14c1.1 0 2-.9 2-2V5c0-1.1-.9-2-2-2zM7.5 15H6v-4.5H4.5V9h3v6zm5.5-1c0 .55-.45 1-1 1H9.5c-.55 0-1-.45-1-1v-4c0-.55.45-1 1-1H12c.55 0 1 .45 1 1v4zm6.5 1h-1.75L16 12.75V15h-1.5V9H16v2.25L17.75 9h1.75l-2.25 3 2.25 3z"/></svg>`,
+  );
+});
+
+Deno.test("Material symbols", async () => {
+  const catalog = materialSymbols();
+
+  assertEquals(
+    await catalog.get("10k"),
+    `<svg xmlns="http://www.w3.org/2000/svg" width="48" height="48" viewBox="0 -960 960 960"><path d="M286-360h50v-240H180v50h106v190Zm120 0h104q14.03 0 23.51-9.49Q543-378.98 543-393v-173q0-14.45-9.49-24.22Q524.03-600 510-600H406q-14.02 0-23.51 9.78Q373-580.45 373-566v173q0 14.02 9.49 23.51Q391.98-360 406-360Zm17-50v-140h70v140h-70Zm157 50h50v-90l84 90h66L668-480l112-120h-66l-84 90v-90h-50v240ZM180-120q-24 0-42-18t-18-42v-600q0-24 18-42t42-18h600q24 0 42 18t18 42v600q0 24-18 42t-42 18H180Zm0-60h600v-600H180v600Zm0-600v600-600Z"/></svg>`,
   );
 });